Fix for decreasing counters
This change makes it so that we no longer try to detect rollovers or restarts.
Instead we just invalidate data in the aggregate when we see a negative
counter delta.

This adds a new row flag called ROW_INVALID. This is used to ensure that a
partial update isn't applied to the trailing edge of the range where there
was a decreasing counter.

Fixes #2.